// src/models/operations/updateimagevisibility.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod/v4-mini";
import { remap as remap$ } from "../../lib/primitives.js";
import { safeParse } from "../../lib/schemas.js";
import { Result as SafeParseResult } from "../../types/fp.js";
import * as types from "../../types/primitives.js";
import * as components from "../components/index.js";
import { SDKValidationError } from "../errors/sdkvalidationerror.js";

export type UpdateImageVisibilityRequest = {
  id: string;
  body: components.UpdateVisibilityRequest;
};

export type UpdateImageVisibilityResponse = {
  httpMeta: components.HTTPMetadata;
  /**
   * Visibility updated
   */
  updateVisibilityResponse?: components.UpdateVisibilityResponse | undefined;
};

/** @internal */
export type UpdateImageVisibilityRequest$Outbound = {
  id: string;
  body: components.UpdateVisibilityRequest$Outbound;
};

/** @internal */
export const UpdateImageVisibilityRequest$outboundSchema: z.ZodMiniType<
  UpdateImageVisibilityRequest$Outbound,
  UpdateImageVisibilityRequest
> = z.object({
  id: z.string(),
  body: components.UpdateVisibilityRequest$outboundSchema,
});

export function updateImageVisibilityRequestToJSON(
  updateImageVisibilityRequest: UpdateImageVisibilityRequest,
): string {
  return JSON.stringify(
    UpdateImageVisibilityRequest$outboundSchema.parse(
      updateImageVisibilityRequest,
    ),
  );
}

/** @internal */
export const UpdateImageVisibilityResponse$inboundSchema: z.ZodMiniType<
  UpdateImageVisibilityResponse,
  unknown
> = z.pipe(
  z.object({
    HttpMeta: components.HTTPMetadata$inboundSchema,
    UpdateVisibilityResponse: types.optional(
      components.UpdateVisibilityResponse$inboundSchema,
    ),
  }),
  z.transform((v) => {
    return remap$(v, {
      "HttpMeta": "httpMeta",
      "UpdateVisibilityResponse": "updateVisibilityResponse",
    });
  }),
);

export function updateImageVisibilityResponseFromJSON(
  jsonString: string,
): SafeParseResult<UpdateImageVisibilityResponse, SDKValidationError> {
  return safeParse(
    jsonString,
    (x) => UpdateImageVisibilityResponse$inboundSchema.parse(JSON.parse(x)),
    `Failed to parse 'UpdateImageVisibilityResponse' from JSON`,
  );
}